chore(comm): Re-aligning naming of compile/source
