+++ module/base/support/phase.init.load.ts
@@ -5,10 +5,10 @@ export const init = {
   key: '@trv:base/load',
   after: ['@trv:base/transpile'],
   action: async () => {
-    const { SourceCodeIndex } = await import('@travetto/boot/src/internal/code');
+    const { SourceIndex } = await import('@travetto/boot/src/internal/source');
     const { AppManifest } = await import('@travetto/base');
 
-    for (const { file } of SourceCodeIndex.findByFolders(AppManifest.source, 'required')) {
+    for (const { file } of SourceIndex.findByFolders(AppManifest.source, 'required')) {
       // Use require vs import on purpose
       require(file); // Scan all files as compiler source root
     }
