+++ module/base/support/phase.init.transpile.ts
@@ -1,3 +1,5 @@
+import { SourceIndex } from '@travetto/boot/src/internal/source';
+
 /**
  * Responsible for transpiling all new code
  */
@@ -5,21 +7,8 @@ export const init = {
   key: '@trv:base/transpile',
   after: '@trv:base/init',
   action: async () => {
-    const { AppCache, EnvUtil } = await import('@travetto/boot');
-    const { SourceCodeIndex } = await import('@travetto/boot/src/internal/code');
-    const { CompileUtil } = await import('@travetto/boot/src/internal/compile');
-
-    if (EnvUtil.isReadonly()) {
-      console.debug('Skipping compilation, in readonly mode');
-      return;
-    }
-
+    const { ModuleManager } = await import('@travetto/boot/src/internal/module');
     const { AppManifest } = await import('@travetto/base');
-
-    for (const x of SourceCodeIndex.findByFolders(AppManifest.source)) {
-      if (!AppCache.hasEntry(x.file)) {
-        CompileUtil.transpile(x.file); // Compile all the desired files
-      }
-    }
+    ModuleManager.transpileAll(SourceIndex.findByFolders(AppManifest.source));
   }
 };
