+++ module/boot/test/code.ts
@@ -1,13 +1,13 @@
 import * as assert from 'assert';
 
 import { Suite, Test } from '@travetto/test';
-import { SourceCodeIndex } from '../src/internal/code';
+import { SourceIndex } from '../src/internal/source';
 
 @Suite()
 class ScanTests {
   @Test()
   testFind() {
-    const files = SourceCodeIndex.find({});
+    const files = SourceIndex.find({});
     assert(files.some(x => x.file.endsWith('test/source.ts')));
   }
 }
