+++ module/cli/src/plugin.ts
@@ -1,4 +1,4 @@
-import { SourceCodeIndex } from '@travetto/boot/src/internal/code';
+import { SourceIndex } from '@travetto/boot/src/internal/source';
 
 import { color } from './color';
 import { BasePlugin } from './plugin-base';
@@ -24,7 +24,7 @@ export class PluginManager {
    */
   static getPluginMapping() {
     const all = new Map<string, string>();
-    for (const { file } of SourceCodeIndex.find({ folder: 'bin', filter: /bin\/cli-/ })) {
+    for (const { file } of SourceIndex.find({ folder: 'bin', filter: /bin\/cli-/ })) {
       all.set(file.replace(/^.*\/bin\/.+?-(.*?)[.][^.]*$/, (_, f) => f), file);
     }
     return all;
