+++ module/compiler/src/transformer.ts
@@ -1,5 +1,6 @@
 import * as ts from 'typescript';
-import { SourceCodeIndex } from '@travetto/boot/src/internal/code';
+
+import { SourceIndex } from '@travetto/boot/src/internal/source';
 
 import {
   NodeTransformer, VisitorFactory, TransformerState, getAllTransformers
@@ -24,7 +25,7 @@ export class TransformerManager {
     }
 
     // Modules
-    const found = SourceCodeIndex.find({ folder: 'support', filter: /\/transformer.*[.]ts/ });
+    const found = SourceIndex.find({ folder: 'support', filter: /\/transformer.*[.]ts/ });
 
     for (const entry of found) { // Exclude based on blacklist
       this.transformers.push(...getAllTransformers(await import(entry.file)));
