+++ module/compiler/support/watch.compiler.ts
@@ -1,7 +1,7 @@
 import { AppManifest, Class, ShutdownManager } from '@travetto/base';
 import { FilePresenceManager, RetargettingProxy } from '@travetto/watch';
 import { FsUtil, PathUtil } from '@travetto/boot';
-import { ModuleUtil } from '@travetto/boot/src/internal/module';
+import { ModuleUtil } from '@travetto/boot/src/internal/module-util';
 
 
 import { Compiler } from '../src/compiler';
