+++ module/transformer/support/phase.init.ts
@@ -1,4 +1,4 @@
-import { SourceUtil } from '@travetto/boot/src/internal/source';
+import { SourceUtil } from '@travetto/boot/src/internal/source-util';
 
 declare const global: { ts: unknown }; // Used for transformers
